requirements.txt
google-generativeai
python-dotenv==1.0.1
pytest==8.1.1
pydantic